+++ CHANGELOG.md
@@ -17,6 +17,11 @@
     Click to see more.
   </summary>
 
+### :feat: New Feature
+
+- `text-color`
+  - [#3812](https://github.com/wix/ricos/pull/3812) support consumer's colors in tiptap (color1,...)
+
 ### :bug: Bug Fix
 
 - `ricos-editor`
@@ -28,6 +33,7 @@
   - [#3806](https://github.com/wix/ricos/pull/3806) RicosContext split from translation handling
   - [#3803](https://github.com/wix/ricos/pull/3803) Modals/Context refactoring
 - `toolbars-v3`
+
   - [#3804](https://github.com/wix/ricos/pull/3804) `tiptapPluginToolbarConfig` and delete atomic nodes action
 
   </details>