+++ CHANGELOG.md
@@ -16,6 +16,11 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+  
+### :house: Internal
+
+- `ricos-editor`
+  - [#3814](https://github.com/wix/ricos/pull/3814) wire draftEditorSettings to tiptap html attributes
 
   </details>
   <hr/>